remove `util` module
Move contents to other modules rather than a catch-all module.
